refactor: rename pkg, more generic implementation
